+++ CollaTime/src/main/resources/templates/common/profile.html
@@ -78,7 +78,7 @@
     let myPicture = "";
     let myColor = "";
 
-    fetch("/mypage/profile")
+    fetch("/common/profile")
         .then(res=>res.json())
         .then(data=>{
             console.log(data)
