+++ CollaTime/src/main/resources/templates/mypage/mypagemain.html
@@ -44,7 +44,7 @@
 <!--    <div id="myPage_participation_project_list">hi</div>-->
     <div th:include="mypage/modifyusermodal.html" id="modifyUser_link"></div>
     <div th:include="mypage/myinquiry.html" id="myInquiry_link"></div>
-    <div th:include="mypage/profile.html" id="myProfile_image"></div>
+    <div th:replace="common/profile.html" id="myProfile_image"></div>
 
 </body>
